+++ src/commands/test/test_scheduler.py
@@ -2,7 +2,7 @@ import multiprocessing
 import signal
 from typing import TYPE_CHECKING, Callable, List, Tuple
 
-from src.commands.test.runner import TestRunner
+from src.commands.test.test_runner import TestRunner
 from src.commands.test.test_live_logger import TestLiveLogger
 from src.commands.test.test_subject_queue import TestSubjectQueue
